requirements.txt
torch==2.2.0         
torchvision==0.17.0    
numpy==1.26.4
scikit-image==0.22.0
opencv-python==4.9.0.80
SimpleITK==2.3.1
tqdm==4.66.1
pyyaml==6.0.1
tensorboard==2.16.2
matplotlib==3.8.2
